feat 세이브 로드 기능을 위한 datamanager, data , test 추가
