+++ Assets/Programing/Jong/Script/SaveTest/Data.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bda0f6f1e94eafe4aafe8b5640fa9eab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
