+++ Assets/Programing/Jong/Script/SaveTest/DataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f6a0c051f952a7145b2032cdf6d5f649
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
