+++ Assets/Programing/Jong/Script/SaveTest/TestChara.cs
@@ -22,41 +22,41 @@ public class TestChara : MonoBehaviour
     // 저장이 필요한 타이밍에 아래와 비슷하게 작성해서 저장 및 불러오기를 해야함 
     private void Start()
     {
-        SaveTest.Instance.LoadGameData(); // 불러오기
+        DataManager.Instance.LoadGameData(); // 불러오기
         Loading();
     }
 
     private void OnApplicationQuit()
     {
         Saving();
-        SaveTest.Instance.SaveGameData(); // 저장하기 
+        DataManager.Instance.SaveGameData(); // 저장하기 
     }
 
     public void Saving() // 씬에 있는 인스턴스로 데이터 전달 
     {
-        SaveTest.Instance.data.isUnlock[0] = firstStage;
-        SaveTest.Instance.data.isUnlock[1] = secondStage;
-        SaveTest.Instance.data.isUnlock[2] = thridStage;
+        DataManager.Instance.data.isUnlock[0] = firstStage;
+        DataManager.Instance.data.isUnlock[1] = secondStage;
+        DataManager.Instance.data.isUnlock[2] = thridStage;
 
 
-        SaveTest.Instance.data.vol = BgmVol;
+        DataManager.Instance.data.vol = BgmVol;
    
 
-        SaveTest.Instance.data.dash = ondash;
-        SaveTest.Instance.data.hover = onhover;
+        DataManager.Instance.data.dash = ondash;
+        DataManager.Instance.data.hover = onhover;
     }
     public void Loading() // json 데이터 파일에서 불러온 데이터를 뽑아옴 
     {
-        firstStage = SaveTest.Instance.data.isUnlock[0];
-        secondStage = SaveTest.Instance.data.isUnlock[1];
-        thridStage = SaveTest.Instance.data.isUnlock[2];
+        firstStage = DataManager.Instance.data.isUnlock[0];
+        secondStage = DataManager.Instance.data.isUnlock[1];
+        thridStage = DataManager.Instance.data.isUnlock[2];
 
 
-        BgmVol = SaveTest.Instance.data.vol;
+        BgmVol = DataManager.Instance.data.vol;
      
 
-        ondash = SaveTest.Instance.data.dash;
-        onhover = SaveTest.Instance.data.hover;
+        ondash = DataManager.Instance.data.dash;
+        onhover = DataManager.Instance.data.hover;
     }
 
 }
